Struct gpiocdev::chip::Chip

pub struct Chip { /* private fields */ }

A GPIO character device.

Implementations

impl Chip

pub fn from_path<P: AsRef<Path>>(p: P) -> Result<Chip>

Constructs a Chip using the given path.

The path must resolve to a valid GPIO character device.

pub fn info(&self) -> Result<Info>

Get the information for the chip.

pub fn name(&self) -> String

Return the name of the chip.

This is based on the filename component of the resolved chip path, not the name from the Info, so it does not involve any system calls.

pub fn path(&self) -> &Path

Return the path of the chip.

pub fn find_line(&self, line: &str) -> Option<Offset>

Find the offset of the named line.

Returns the first matching line.

pub fn line_info(&self, offset: u32) -> Result<Info>

Get the information for a line on the chip.

pub fn watch_line_info(&self, offset: u32) -> Result<Info>

Add a watch for changes to the publicly available information on a line.

This is a null operation if there is already a watch on the line.

pub fn unwatch_line_info(&self, offset: u32) -> Result<()>

Remove a watch for changes to the publicly available information on a line.

This is a null operation if there is no existing watch on the line.

pub fn has_line_info_change_event(&self) -> Result<bool>

Check if the request has at least one info change event available to read.

pub fn wait_line_info_change_event(&self, timeout: Duration) -> Result<bool>

Wait for an info change event to be available.

pub fn read_line_info_change_event(&self) -> Result<InfoChangeEvent>

Read a single line info change event from the chip.

Will block until an edge event is available.

pub fn info_change_events(&self) -> Result<InfoChangeIterator<'_>>

An iterator for info change events from the chip.

pub fn detect_abi_version(&self) -> Result<AbiVersion>

Detect the most recent uAPI ABI supported by the library for the chip.

pub fn supports_abi_version(&self, abiv: AbiVersion) -> Result<()>

Check if the platform and library support a specific ABI version.

pub fn using_abi_version(&mut self, abiv: AbiVersion) -> &mut Self

Set the ABI version to use for subsequent operations.