Module gpiocdev::line

Types specific to lines.

Structs

Config

The configuration settings for a single line.

EdgeEvent

The details of an edge detected on an input line.

Info

The publicly available information for a line.

InfoChangeEvent

The details of a change to the Info for a line.

Values

A collection of line values.

Enums

Bias

The bias settings for a line.

Direction

The direction of a line.

Drive

The drive policy settings for an output line.

EdgeDetection

The edge detection options for an input line.

EdgeKind

The cause of an EdgeEvent.

EventClock

The available clock sources for EdgeEvent timestamps.

InfoChangeKind

The cause of a InfoChangeEvent

Value

The logical level of a line.

Type Definitions

Offset

An identifier for a line on a particular chip.

Offsets

A collection of line offsets.