Module gpiocdev::request

Types and functions related to requesting lines.

The Builder creates the Request which can access and manipulate sets of lines, and return edge events via the EdgeEventBuffer.

The line configuration can be provided to the Builder using either direct mutators, such as as_input and with_edge_detection, or via a Config.

To request and read a basic input line:

use gpiocdev::request::Request;

let l3 = Request::builder()
    .on_chip("/dev/gpiochip0")
    .with_line(3)
    .request()?;
let value = l3.value(3)?;

Structs

Builder

A builder of line requests.

Config

The configuration for a request for one or more lines.

EdgeEventBuffer

A user space buffer for reading edge events in bulk from a Request.

Request

An active request of a set of lines.