Struct gpiocdev::request::Builder

pub struct Builder { /* private fields */ }

A builder of line requests.

Most mutators operate on the request configuration as per the Config.

To request and read a basic input line:

use gpiocdev::request::Request;

let l3 = Request::builder()
    .on_chip("/dev/gpiochip0")
    .with_line(3)
    .request()?;
let value = l3.value(3)?;

Or several lines:

use gpiocdev::request::Request;
use gpiocdev::line::Values;

let offsets = &[3,5];
let ll = Request::builder()
    .on_chip("/dev/gpiochip0")
    .with_lines(offsets)
    .request()?;
let mut values = Values::from_offsets(offsets);
ll.values(&mut values)?;

Alternatively, more complex configurations can be built separately and provided to the Builder:

use gpiocdev::request::{Config, Request};
use gpiocdev::line::Value;

let mut cfg = Config::default();
cfg.with_line(5).as_output(Value::Active);
let l5 = Request::from_config(cfg)
    .on_chip("/dev/gpiochip0")
    .request()?;
l5.set_value(5,Value::Inactive)?;

Implementations

impl Builder

pub fn request(&self) -> Result<Request>

Perform the request.

Sends the request to the kernel using the appropriate uAPI call.

On success returns the Request that provides access to the requested lines.

This is the terminal operation for the Builder.

pub fn with_config(&mut self, cfg: Config) -> &mut Self

Replace the request configuration with the new one provided.

pub fn config(&self) -> Config

A snapshot of the current request configuration.

Can be applied to other requests with with_config.

pub fn with_consumer<N: Into<String>>(&mut self, consumer: N) -> &mut Self

Specify the consumer label to be applied to the request, and so to all lines in the request.

If not specified, a label “gpiocdev-pPID” is applied by request, where PID is the process id of the application.

pub fn with_kernel_event_buffer_size(
&mut self,
event_buffer_size: u32
) -> &mut Self

Set the event buffer size for edge events buffered in the kernel.

This method is only required in unusual circumstances.

The size provides a hint to the kernel as to the size of the buffer to store edge events until they are read from the Request. The default size is 16 times the number of lines in the request. The default size is typically sufficient unless the hardware may generate bursts of events faster than the application can service them.

Altering the buffer size does NOT affect latency. Buffering is provided in the kernel to reduce the likelyhood of event loss when user space is slow servicing events. In all cases the events are provided to user space as fast as user space allows.

pub fn with_user_event_buffer_size(
&mut self,
event_buffer_size: usize
) -> &mut Self

Set the event buffer size for edge events buffered in user space.

This method is only required in unusual circumstances.

The size defines the number of events that may be buffered in user space by the Request.edge_events iterator. The user space buffer is a performance optimisation to reduce the number of system calls required to read events.

Altering the buffer size does NOT affect latency. In all cases the events are provided to user space as fast as user space allows.

pub fn using_abi_version(&mut self, abiv: AbiVersion) -> &mut Self

Select the ABI version to use when requesting the lines and for subsequent operations.

This is typically only required where an application supports both uAPI versions but the kernel may only support uAPI v1.

The version to use can be determined by detect_abi_version.

pub fn on_chip<P: Into<PathBuf>>(&mut self, path: P) -> &mut Self

Set the chip from which to request lines.

This applies to all lines in the request. It is not possible to request lines from different chips in the same request.

The chip is identified by a path which must resolve to a GPIO character device.

pub fn as_input(&mut self) -> &mut Self

Set the selected lines to input.

This is a short form of with_direction(Input).

This is the default direction setting.

pub fn as_output(&mut self, value: Value) -> &mut Self

Set the selected lines to output with the given value.

This is a short form of with_direction(Output) and with_value(value) that allows the direction and value to be set in the one call.

pub fn as_active_low(&mut self) -> &mut Self

Set the selected lines to active low.

pub fn as_active_high(&mut self) -> &mut Self

Set the selected lines to active high.

This is the default active level setting.

pub fn with_bias<B: Into<Option<Bias>>>(&mut self, bias: B) -> &mut Self

Set the bias setting for the selected lines.

pub fn with_debounce_period(&mut self, period: Duration) -> &mut Self

Set the debounce period for the selected lines.

A value of zero means no debounce.

Implicitly sets the selected lines as inputs, if they weren’t already, and removes any output specific settings.

pub fn with_direction(&mut self, direction: Direction) -> &mut Self

Set the direction of the selected lines.

Setting to input removes any output specific settings.

Setting to output removes any input specific settings.

pub fn with_drive(&mut self, drive: Drive) -> &mut Self

Set the drive setting for the selected lines.

Implicitly sets the selected lines as outputs, if they weren’t already, and removes any input specific settings.

pub fn with_edge_detection<E: Into<Option<EdgeDetection>>>(
&mut self,
edge: E
) -> &mut Self

Set the edge detection for the selected lines.

Implicitly sets the lines as inputs and removes any output specific settings.

pub fn with_event_clock(&mut self, event_clock: EventClock) -> &mut Self

Set the clock source for edge events on the selected lines.

pub fn with_line(&mut self, offset: Offset) -> &mut Self

Add a line to the request.

Note that all configuration mutators applied subsequently only apply to this line.

pub fn without_line(&mut self, offset: Offset) -> &mut Self

Remove a lines from the request.

pub fn with_lines(&mut self, offsets: &[Offset]) -> &mut Self

Add a set of lines to the request.

Note that all configuration mutators applied subsequently only apply to this subset of lines.

pub fn without_lines(&mut self, offsets: &[Offset]) -> &mut Self

Remove a set of lines from the request.

pub fn with_value(&mut self, value: Value) -> &mut Self

Set the value of the selected lines.

pub fn from_line_config(&mut self, cfg: &Config) -> &mut Self

Apply the configuration based on a snapshot from a single line.